% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckPoly.R
\name{CheckPoly}
\alias{CheckPoly}
\title{Check your SpatialPolygonDataFrame's compatibility with PolygonMap() function.}
\usage{
CheckPoly(shp)
}
\arguments{
\item{shp}{A SpatialPolygonsDataFrame}
}
\value{
TRUE, or function will STOP with print statements about what requirements were not met.
}
\description{
Checks the spatial object you intend to use with the PolygonMap() function
for valid geometry and projection
}
\examples{
check_wmapshp(a_SpatialPolygonsDataFrame)
}

