% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExtent.R
\name{GetExtent}
\alias{GetExtent}
\title{GetExtent}
\usage{
GetExtent(map_object, return_table = F)
}
\arguments{
\item{map_object}{A ggplot object}

\item{return_table}{Boolean; Default=F. If T, returns a table where
xbounds1 is the xmin, xbounds2 is the xmax, ybounds1 is the ymin,
and ybounds2 is the ymax.}
}
\value{
Either a named list where the results has an $xbounds and
      $ybounds named list (with a min/max in each), or a table (see
      return_table parameter).
}
\description{
Gets the x and y coordinates of the full extent of a ggplot.
}
\examples{

GetExtent(map_obj)

}

