% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLegend.R
\name{GetLegend}
\alias{GetLegend}
\title{GetLegend}
\usage{
GetLegend(ggplot)
}
\arguments{
\item{ggplot}{A ggplot object}
}
\value{
A Grob of the ggplot's legend if one exists,
        or a string of text "Map has no legend."
}
\description{
Description of function here
}
\examples{

GetLegend(your_map)

}

