% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistogramColorstats.R
\name{HistogramColorstats}
\alias{HistogramColorstats}
\title{Histogram of data using color scheme}
\usage{
HistogramColorstats(datavector, color_ramp = "grey17", minimum = NULL,
  maximum = NULL, color_value_breaks = NULL, dist_stats = NULL,
  mean_color = "red", quantile_color = "black", title = "")
}
\arguments{
\item{datavector}{vector of color data}

\item{color_ramp}{color scheme for the histogram}

\item{minimum}{min value to display in histogram}

\item{maximum}{max value to display in histogram}

\item{color_value_breaks}{values to break colors at}

\item{dist_stats}{which distribution to use}

\item{mean_color}{color to use at mean}

\item{quantile_color}{color to use for quantiles}
}
\value{
ggplot histogram
}
\description{
This histogram function creates and returns a ggplot object with
a color scale dependent on the X value (default as all colors being grey)
as well as the possibility for vertical lines that describe different
properties of the distribution (mean, median, etc) added on top of the graph.
}
\examples{
histogram_colorstats(rnorm(100))
}

