% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntensityPalette.R
\name{IntensityPalette}
\alias{IntensityPalette}
\title{Cubehelix Color Pallette with Increasing Intensity}
\usage{
IntensityPalette(start, r, gamma, hue_list = c(0.5, 1, 3))
}
\arguments{
\item{start}{numeric start-point for cubehelix color ramp}

\item{r}{number of rotations through cubehelix color ramp}

\item{gamma}{numeric; how light or dark you want it to be 1: normal, higher:darker, lower: lighter}

\item{hue}{numeric;how bright you want the colors to be 1: normal, higher: brighter, lower: more demure}

\item{color_value_breaks}{values to break colors at}

\item{dist_stats}{which distribution to use}
}
\value{
an adapted cubehelix color pallette, increasing in intensity.
}
\description{
This function returns a cubehelix color pallette 
that increases in intensity as it moves from light to dark.
}
\examples{
IntensityPalette(start=0.5,r=.5,gamma=1)
}

