% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotColors.R
\name{PlotColors}
\alias{PlotColors}
\title{Plot colors of a color pallette}
\usage{
PlotColors(color_list, color_list_name = NULL, n = NULL, base_size = 11,
  listcolors = F, labelcolors = T)
}
\arguments{
\item{color_list}{list or vector of valid R colors}

\item{color_list_name}{name of color scheme, optional, default is NULL.}
}
\value{
ggplot of color swatches and optional name of color pallette
}
\description{
This function creates colored swatches of a list
of colors for easy visualization.
}
\examples{
PlotColors(c("red","yellow","blue"),"primary colors")
}

