% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShptoSPDF.R
\name{ShptoSPDF}
\alias{ShptoSPDF}
\title{ShptoSPDF}
\usage{
ShptoSPDF(shp_dir, shp_layer, id_field, all_data = T)
}
\arguments{
\item{shp_dir}{The directory where your shapefile is}

\item{shp_layer}{The name of the shapefile (ex: "census_tracts"), with no .shp or any other extension}

\item{id_field}{Character string of the column that is the primary key (what defines polygons as being
different from one another) in the shapefile. If a field that is not a primary key is provided, the polygons
will be collapsed such that there will be 1 polygon object (possibly multi-part) for each of the primary keys
you have identified, but the rest of the data table will no longer attach to the shapefile.}

\item{all_data}{Logical; Whether you want the rest of the data attributes (other than the id_field provided)
to remain in the @data slot of the SpatialPolygonsDataFrame.}
}
\value{
A SpatialPolygonsDataFrame with a data.table
}
\description{
Brings in a shapefile (from a directory on the computer) as a SpatialPolygonsDataFrame
with a data.table as the data attributes.
}
\examples{
shp_to_Rpolygons("C:/Users/username/Documents/spatial_data/","shapfile_name","fipscode")
}

