% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackLegends.R
\name{StackLegends}
\alias{StackLegends}
\title{StackLegends}
\usage{
StackLegends(legends)
}
\arguments{
\item{legends}{List of maps/ggplots with legends. If you give the function
an object without a legend in this list, it will error out, noting that 
one of the maps does not contain a legend. List order is significant: 
first map's legend will be on the bottom, second map's legend above it, etc.}
}
\value{
A ggplot object with legends vertically stacked.
}
\description{
Stacks legends from different plots into 1 ggplot object.
}

