% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackMaps.R
\name{StackMaps}
\alias{StackMaps}
\title{Stack Maps}
\usage{
StackMaps(maps, xbounds = NULL, ybounds = NULL, font_family = "serif",
  map_title = NULL, map_subtitle = NULL, map_title_justification = 0.5,
  map_title_font_size = 20, map_title_font_face = "plain")
}
\arguments{
\item{maps}{A list of ggplot map objects all within the same coordinate reference system. Order of
the list is significant- the first object will be on the bottom, second layered above it, etc.}

\item{xbounds}{Default=NULL. Takes a 2-item list of xmin and xmax of the area you want to plot.
Leaving this parameter as null will default to having the bounds ofthe map be the 
minimum and maximum x coordinates present in any of the map layers.}

\item{ybounds}{Default=NULL. Takes a 2-item list of xmin and xmax of the area you want to plot.
Leaving this parameter as null will default to having the bounds ofthe map be the 
minimum and maximum y coordinates present in any of the map layers.}

\item{font_family}{The name of the font family you want to use for the title text
on the plot. Default is 'serif'.}

\item{map_title}{String; the title of your map.}

\item{map_subtitle}{String; Default=NULL}

\item{map_title_justification}{Numeric from 0-left to 1-right. Default is centeR-.5.}

\item{map_title_font_size}{Numeric, default is 20. Note that subtitle size will be 
scaled to .8 of this parameter.}

\item{map_title_font_face}{Default="plain".}
}
\value{
A ggplot object with maps stacked
}
\description{
Stacks ggplot objects over one another to generate 1 multi-layered
     map object.
}
\examples{

StackMaps(list_of_maps)

}

