% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewPal.R
\name{ViewPal}
\alias{ViewPal}
\title{View a Woodson Pallette color scheme.}
\usage{
ViewPal(color = NULL, n = NULL, base_size = 11, listcolors = F,
  n_per_page = 5, labelcolors = T)
}
\arguments{
\item{color}{the name of the color scheme}
}
\value{
plots of one or all Woodson pallette schemes
}
\description{
This function shows a woodson pallette color scheme when the name is entered. 
 When no name of a wpal() color scheme is entered, all woodson pallette shemes are shown.
}
\examples{
ViewPal()
}

