% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplayout.R
\name{vplayout}
\alias{vplayout}
\title{Grid viewport function wrapper.}
\usage{
vplayout(x, y, ...)
}
\arguments{
\item{x}{float > 0}

\item{y}{float > 0}

\item{...}{further argumenst to be passed to the grid::viewport function}
}
\value{
An R object of class viewport.
}
\description{
A wrapper for the viewport function that overwrites the original
parameters.
}

