% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpal.R
\name{wpal}
\alias{wpal}
\title{Map Colors Palettes}
\usage{
wpal(color = NULL, n = NULL, noblack = FALSE)
}
\arguments{
\item{color}{Name of the color ramp. Default is NULL, which will return the whole list,
so you should probably enter in a color name! Use the funciton view_pal() to see your options.}

\item{n}{How many colors are returned from the ramp. Default is NULL, which will provide the number
of colors generated when the ramp is created.}

\item{noblack}{Logical; Defines whether the color black is to be excluded from your color ramp. Default value is FALSE.}
}
\value{
A list of color values.
}
\description{
A suite of color ramps designed for beautiful chloropleth maps and other
data visualizations, especially where value differentiation over a wide range is desired.
}
\examples{
MapColors("cool_toned")
}

